Add new item data
Copy over old TMs, etc.

+++ include/item.h
@@ -9,7 +9,7 @@ struct Item
 {
     u8 name[ITEM_NAME_LENGTH];
     u16 itemId;
-    u16 price;
+    u32 price;
     u8 holdEffect;
     u8 holdEffectParam;
     const u8 *description;
